# doc/api.yaml
openapi: 3.0.0
info:
  title: BTD Project
  description: Sample NodeJS BTD spec
  version: '1.0.0'
servers:
  - url: http://localhost:3000
    description: dev server
security:
  - BearerAuth: []
...
paths:
...
  /api/users:
    get:
      tags:
        - user
      summary: Get all users
      security:
        - BearerAuth: []
      description: Get all users
      responses:
        '200':
          description: Gets list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
                
    post:
      tags:
        - user
      summary: Create new user
      security:
        - BearerAuth: []
      description: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
...
components:
  schemas:
...
    Body:
      type: object
      properties:
        name:
          type: string
          minLength: 10
          maxLength: 100
          example: Jonh Doe
        email:
          type: string
          example: jonh-doe@example.com
        password:
          type: string
          example: '123456'
...
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User" 
    User: 
      type: object
      required:
        - name
        - role
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: Jonh Doe
        role:
          type: string
          enum:
            - Admin
            - User
          example: User
        isActive:
          type: boolean
          example: false
        email:
          type: string
          example: jonh-doe@example.com
        password:
          type: string
          example: '123456'
...
    ServerError:
      type: object
      properties:
        status: 
          type: integer
          enum:
           - 500
        error:
          type: string
          example: 'internal-server-error'
        message:
          type: string
          example: 'Something went wrong'
        extraDetails:
          type: string
          example: 'Extra error details'
  responses:
...
    UnauthorizedError:
      description: Access token is missing or invalid
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT